Small tweaks for truing stand slider dimensions.
Slides rather stiff when printed with 0.03 slack. Nuts were
very loose at 0.48.

This script has been tweaked/cleaned up from the original (full of
magic numbers), so these modifications are expected. Not going
to bother doing another print until another friend asks for a
stand - going to assume this works well enough, and if not, a bit
of heat or acetone will set things in properly.