Adding dimensions for a 2.5" Ridgid shop vac hose.
Note that the print for the inlet was a blonde one too loose, so had to
paint the inside with a thin layer of superglue, then baking soda, and then
sand a smidge to even it out.

Also saving the old 1.875" Ridgid measurements to their own file.

Really wish I had some slick way for doing dimension inheritance now... Hrm.